@cc-code/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.24.3
jieba==0.42.1
wordcloud==1.9.2
matplotlib==3.8.2
flask==3.0.0
Pillow==10.1.0
lxml==4.9.3
fake-useragent==1.4.0
webdriver-manager==4.0.1
scikit-learn==1.3.2
textblob==0.17.1
pyecharts==1.9.1
snownlp==0.12.3
aiohttp==3.9.1
asyncio
schedule==1.2.0